+++ src/index.js
@@ -1,4 +1,5 @@
 export { encryptValue } from './crypto/encrypt.js';
 export { decryptValue } from './crypto/decrypt.js';
 export { processConfig } from './utils/config.js';
+export { serializePath } from './utils/path.js';
 export { listSupportedAlgorithms as getSupportedAlgorithms } from './crypto/utils.js';